Create BusyBox's sh symlink before starting e2e tests
