+++ test/e2e/_common.js
@@ -3,7 +3,6 @@
  * @license MIT
  */
 
-import fs from "node:fs";
 import path from "node:path";
 import process from "node:process";
 
@@ -32,7 +31,7 @@ export function getTestArgs() {
 export function getTestFn(shell) {
   try {
     if (!isCI && typeof shell === "string") {
-      locate(shell);
+      which.sync(shell, { path: process.env.PATH || process.env.Path });
     }
 
     return test;
@@ -56,44 +55,11 @@ export function getTestShells() {
     if (constants.isMacOS) {
       systemShells.splice(busyboxIndex, 1);
     } else {
-      systemShells[busyboxIndex] = createBusyBoxSh();
+      const root = path.resolve(import.meta.dirname, "..", "..");
+      const temp = path.resolve(root, ".temp");
+      systemShells[busyboxIndex] = path.resolve(temp, "busybox", "sh");
     }
   }
 
   return [false, ...systemShells];
-}
-
-/**
- * Create a link file called "sh" to BusyBox so that it can be used as a shell.
- *
- * @returns {string} The location of the sh file linking to BusyBox.
- */
-export function createBusyBoxSh() {
-  const root = path.resolve(import.meta.dirname, "..", "..");
-  const temp = path.resolve(root, ".temp", "busybox");
-  if (!fs.existsSync(temp)) {
-    fs.mkdirSync(temp, { recursive: true });
-  }
-
-  const busyboxSh = path.resolve(temp, "sh");
-  if (!fs.existsSync(busyboxSh)) {
-    try {
-      const busybox = locate(constants.binBusyBox);
-      fs.symlinkSync(busybox, busyboxSh);
-    } catch {
-      return constants.binBusyBox;
-    }
-  }
-
-  return busyboxSh;
-}
-
-/**
- * Obtain the location of a binary on the system by name.
- *
- * @param {string} binary The name of the binary to location.
- * @returns {string} The path to the binary.
- */
-function locate(binary) {
-  return which.sync(binary, { path: process.env.PATH || process.env.Path });
 }
